+++ GameRefactoring/Assets/Scenes/ForestRunScene.unity
@@ -384,7 +384,8 @@ PrefabInstance:
       propertyPath: attackCooldown
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 5286118462819963944, guid: 0afd4ab769ba45a4cb51ef6ecff10292, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 0afd4ab769ba45a4cb51ef6ecff10292, type: 3}
 --- !u!1001 &6659038790531184164
 PrefabInstance:
