+++ GameRefactoring/Assets/Sprites/Interface.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e1922489f48f954abfee89e90e2c3ca
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
