+++ GameRefactoring/Assets/Sprites/Interface/hp_mana.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 061b643715f345744a09cb738a9e38f8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
