+++ GameRefactoring/Assets/Sprites/PickUps_and_GameObjects/Coin.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d04ac6bffd1910d47ad138e45c32be91
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
