+++ GameRefactoring/Assets/Sprites/PickUps_and_GameObjects/Shield.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a0371e705a4380c46b8415b2cd768aea
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
