+++ GameRefactoring/Assets/Sprites/Player/PlayerSkins/Idle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 01c80924df35bce43be5d5d1d1333781
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
