+++ GameRefactoring/Assets/Sprites/TownParts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fd19b71326c382c4bb3b5956f6883b43
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
