+++ GameRefactoring/Assets/Sprites/TownParts/NPC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab251c9708e96bf46905538951a0b2ca
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
