+++ GameRefactoring/Assets/Sprites/TownParts/NPC/Tailor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab64261bb26612345b661a52b7a50a15
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
