+++ GameRefactoring/Assets/Sprites/TownParts/NPC/Tailor/Idle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06955a580a648744caa0402505f7f8b1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
