+++ GameRefactoring/Assets/Sprites/TownParts/Portal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d950ab003d114a42b9aaa65a389d8cc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
